+++ example/indieauth.php
@@ -21,14 +21,27 @@ require_once dirname(__DIR__).'/vendor/autoload.php';
 use fkooman\Rest\Service;
 use fkooman\Http\Request;
 use fkooman\Rest\Plugin\IndieAuth\IndieAuthAuthentication;
-use fkooman\Rest\Plugin\UserInfo;
+use fkooman\Rest\Plugin\IndieAuth\IndieInfo;
+use GuzzleHttp\Client;
 
 try {
     $service = new Service();
-    $service->setDefaultRoute('/welcome');
 
     // use discovery by default, fall back to IndieCert (https://indiecert.net/auth)
     $indieAuth = new IndieAuthAuthentication();
+    
+    // in case no session is active yet, redirect to '/'
+    $indieAuth->setUnauthorizedRedirectUri('/');
+
+    $client = new Client(
+        array(
+            'defaults' => array(
+                'verify' => false,
+                'timeout' => 10
+            )
+        )
+    );
+    $indieAuth->setClient($client);
 
     // use discovery by default, fall back to IndieAuth (https://indieauth.com/auth)
     //$indieAuth = new IndieAuthAuthentication('https://indieauth.com/auth');
@@ -39,13 +52,22 @@ try {
     $service->registerOnMatchPlugin($indieAuth);
 
     $service->get(
-        '/welcome',
+        '/',
         function (Request $request) {
             // Show Sign In form;  POST to 'indieauth/auth' endpoint which is registered by the IndieAuth plugin
-            return '<html><head></head><body><h1>Sign In</h1><form method="post" action="indieauth/auth">https://<input type="text" name="me" placeholder="example.org" required><input type="hidden" name="redirect_to" value="/success"><input type="submit" value="Sign In"></form></body></html>';
+            $output = '<html><head></style></head><body><h1>Sign In</h1>';
+
+            // without requesting access token (no scope)
+            $output .= '<h2>Authentication Only</h2><form method="post" action="indieauth/auth">https://<input type="text" name="me" placeholder="example.org" required><input type="hidden" name="redirect_to" value="/success"><input type="submit" value="Sign In"></form>';
+
+            // with requesting access token (scope)
+            $output .= '<h2>Authentication &amp; Token</h2><form method="post" action="indieauth/auth">https://<input type="text" name="me" placeholder="example.org" required><input type="text" name="scope" value="post"><input type="hidden" name="redirect_to" value="/success"><input type="submit" value="Sign In"></form>';
+
+            $output .= '</body></html>';
+            return $output;
         },
         array(
-            // To view the "welcome" page, no authentication is required
+            // To view the index page, no authentication is required
             'skipPlugins' => array(
                 'fkooman\Rest\Plugin\IndieAuth\IndieAuthAuthentication'
              )
@@ -55,10 +77,10 @@ try {
     // To view the "success" page, authentication is required
     $service->get(
         '/success',
-        function (UserInfo $u) {
+        function (IndieInfo $u) {
             return sprintf(
-                '<html><head></head><body><h1>Hello</h1><p>Hello %s</p><p><a href="indieauth/logout">logout</a></p></body></html>',
-                $u->getUserId()
+                '<html><head></head><body><h1>Hello</h1><table><tr><th>User ID</th><td>%s</td></tr><th>Access Token</th><td>%s</td></tr><tr><th>Scope</th><td>%s</td></tr></table><p><a href="indieauth/logout">logout</a></p></body></html>',
+                $u->getUserId(), $u->getAccessToken(), $u->getScope()
             );
         }
     );